// Vrchat OSC/OSC Vrchat server/src/models/enums.ts
export const enum ChessIndexName {
  Empty_1 = "Empty_1",
  Rook_White_1 = "Rook_White_1",
  Knight_White_1 = "Knight_White_1",
  Bishop_White_1 = "Bishop_White_1",
  Queen_White_1 = "Queen_White_1",
  King_White_1 = "King_White_1",
  Bishop_White_2 = "Bishop_White_2",
  Knight_White_2 = "Knight_White_2",
  Rook_White_2 = "Rook_White_2",
  Pawn_White_1 = "Pawn_White_1",
  Pawn_White_2 = "Pawn_White_2",
  Pawn_White_3 = "Pawn_White_3",
  Pawn_White_4 = "Pawn_White_4",
  Pawn_White_5 = "Pawn_White_5",
  Pawn_White_6 = "Pawn_White_6",
  Pawn_White_7 = "Pawn_White_7",
  Pawn_White_8 = "Pawn_White_8",
  Pawn_Black_1 = "Pawn_Black_1",
  Pawn_Black_2 = "Pawn_Black_2",
  Pawn_Black_3 = "Pawn_Black_3",
  Pawn_Black_4 = "Pawn_Black_4",
  Pawn_Black_5 = "Pawn_Black_5",
  Pawn_Black_6 = "Pawn_Black_6",
  Pawn_Black_7 = "Pawn_Black_7",
  Pawn_Black_8 = "Pawn_Black_8",
  Rook_Black_1 = "Rook_Black_1",
  Knight_Black_1 = "Knight_Black_1",
  Bishop_Black_1 = "Bishop_Black_1",
  Queen_Black_1 = "Queen_Black_1",
  King_Black_1 = "King_Black_1",
  Bishop_Black_2 = "Bishop_Black_2",
  Knight_Black_2 = "Knight_Black_2",
  Rook_Black_2 = "Rook_Black_2",
};